some SQL syntax does not pass a table prefix (atom or string)
in the case of `create_name`, then third param may be a string or atom
prefix ... and in those cases should be treated as a prefix.

however syntax such as `VALUES (...)` will pass in the values list
information as the third tuple. this was causing spurious exceptions
to be thrown, rather than creating the correct table name